requirements.txt
numpy>=2.4,<3
opencv-python-headless>=4.13,<5
Pillow>=12.2,<13
PyAudio>=0.2.14,<0.3
PyInstaller>=6.19,<7
pynput>=1.8,<2
pyvirtualcam>=0.15,<0.16
typing-extensions>=4.15,<5
